move project types to a separate types file
