+++ cypress-tests/cypress/support/types/project.types.ts
@@ -1,0 +1,10 @@
+export type ProjectIdentifierV2 = {
+  id?: string;
+  namespace?: string;
+  slug: string;
+};
+
+export interface NewProjectV2Body extends ProjectIdentifierV2 {
+  name: string;
+  visibility?: "public" | "private";
+}
